Rename and move classes
