+++ executor-service/src/main/java/executor/service/service/step/Impl/StepExecutionClickCss.java
@@ -1,12 +1,12 @@
-package executor.service.service.Impl;
+package executor.service.service.step.Impl;
 
 import executor.service.model.Step;
-import executor.service.service.StepExecutionClickCss;
+import executor.service.service.step.StepExecution;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-public class StepExecutionClickCssImpl implements StepExecutionClickCss {
+public class StepExecutionClickCss implements StepExecution {
     @Override
     public String getStepAction() {
         return "clickCss";
