+++ executor-service/src/main/java/executor/service/service/step/StepExecution.java
@@ -1,9 +1,9 @@
-package executor.service.service;
+package executor.service.service.step;
 
 import executor.service.model.Step;
 import org.openqa.selenium.WebDriver;
 
-public interface StepExecutionClickCss {
+public interface StepExecution {
     String getStepAction();
     void step(WebDriver webDriver, Step step);
 }
